refactor: ♻️ use nestjs/config in config lib
