+++ .env.sample
@@ -1,8 +1,11 @@
-GATEWAY_HOST=localhost
+GATEWAY_HOST=http://localhost
 GATEWAY_PORT=3333
 
-USER_HOST=localhost
+USER_HOST=http://localhost
 USER_PORT=15001
 
-TASK_HOST=localhost
+TASK_HOST=http://localhost
 TASK_PORT=15002
+
+DATABASE_USER=test
+DATABASE_PASSWORD=test
