--- libs/application-config/project.json
@@ -1,9 +1,0 @@
-{
-  "name": "applications-config",
-  "$schema": "../../node_modules/nx/schemas/project-schema.json",
-  "sourceRoot": "libs/application-config/src",
-  "projectType": "library",
-  "tags": [],
-  "// targets": "to see all targets run: nx show project applications-config --web",
-  "targets": {}
-}
