--- libs/application-config/src/index.ts
@@ -1,2 +1,0 @@
-export * from './lib/applications-config';
-export * from './lib/subgraphs-config';
