--- libs/application-config/src/lib/subgraphs-config.ts
@@ -1,10 +1,0 @@
-import { ServiceEndpointDefinition } from '@apollo/gateway';
-
-import { userSubGraph } from './applications-config';
-
-export const subgraphsConfig: ServiceEndpointDefinition[] = [
-  {
-    name: userSubGraph.name,
-    url: `${userSubGraph.host}:${userSubGraph.port}/graphql`,
-  },
-];
