+++ libs/config/jest.config.ts
@@ -1,10 +1,10 @@
 export default {
-  displayName: 'applications-config',
+  displayName: 'config',
   preset: '../../jest.preset.js',
   testEnvironment: 'node',
   transform: {
     '^.+\\.[tj]s$': ['ts-jest', { tsconfig: '<rootDir>/tsconfig.spec.json' }],
   },
   moduleFileExtensions: ['ts', 'js', 'html'],
-  coverageDirectory: '../../coverage/libs/application-config',
+  coverageDirectory: '../../coverage/libs/config',
 };
