+++ libs/config/src/index.ts
@@ -1,0 +1,4 @@
+export * from './lib/applications.config';
+export * from './lib/database.config';
+
+export * from './lib/subgraphs.config';
