+++ libs/config/tsconfig.spec.json
