+++ package.json
@@ -17,6 +17,7 @@
     "@apollo/subgraph": "^2.5.5",
     "@nestjs/apollo": "^12.2.1",
     "@nestjs/common": "^10.4.7",
+    "@nestjs/config": "^3.3.0",
     "@nestjs/core": "^10.4.7",
     "@nestjs/graphql": "^12.2.1",
     "@nestjs/platform-express": "^10.4.7",
